+++ packages/nimble-components/src/table-column/mixins/tests/fractional-width-column.spec.ts
@@ -22,6 +22,7 @@ const columnName = uniqueElementName();
     name: columnName
 })
 class TestTableColumn extends mixinFractionalWidthColumnAPI(TableColumn) {
+    public groupHeaderViewTag = '';
     public constructor() {
         super({
             cellRecordFieldNames: [],